TAF-15 Move abstract request to common module
