+++ skyxplore-api-test/src/test/java/com/mentoring/endpoints/User.java
@@ -1,5 +1,6 @@
 package com.mentoring.endpoints;
 
+import com.mentoring.api.AbstractRequest;
 import com.mentoring.model.requestbody.UserRequest;
 import io.restassured.response.Response;
 import lombok.extern.slf4j.Slf4j;
